refactor(entity): 消除 Entity 层冗余 - 提取 BaseTrade
## 变更内容

根据 `docs/entity-redundancy-analysis.md` 分析报告，提取 Trade 和 BacktestTrade 的共享字段到 BaseTrade。

### 新增文件

- BaseTrade.java: @MappedSuperclass 基类，包含共享字段
  - tradeType, symbol, price, quantity, amount, tradeTime, createdAt
- TradeStatus.java: 从 Trade 中提取的枚举

### 修改的文件

- Trade.java: 继承 BaseTrade，使用 TradeStatus 枚举
- BacktestTrade.java: 继承 BaseTrade
- User.java: 修复（保持 UserStatus 内部枚举）

### 质量检查结果

- mvn clean compile: BUILD SUCCESS
- checkstyle:check: 0 violations
- quality-check.sh: 8/8 全绿

## 后续工作

剩余冗余问题（内部枚举提取、手写 Builder 替换）将在后续迭代中处理。

Closes #398
