+++ koduck-backend/src/main/java/com/koduck/entity/Trade.java
@@ -1,8 +1,5 @@
 package com.koduck.entity;
 
-import java.math.BigDecimal;
-import java.time.LocalDateTime;
-
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
@@ -13,16 +10,15 @@ import jakarta.persistence.Id;
 import jakarta.persistence.Index;
 import jakarta.persistence.Table;
 
-import org.hibernate.annotations.CreationTimestamp;
-
-import com.koduck.entity.enums.TradeType;
+import com.koduck.entity.enums.TradeStatus;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import lombok.experimental.SuperBuilder;
 
 /**
  * Trade entity.
@@ -38,10 +34,11 @@ import lombok.Setter;
        }
 )
 @Data
-@Builder
+@SuperBuilder
 @NoArgsConstructor
 @AllArgsConstructor
-public class Trade {
+@EqualsAndHashCode(callSuper = true)
+public class Trade extends BaseTrade {
 
     /** The ID. */
     @Id
@@ -57,66 +54,16 @@ public class Trade {
     @Column(name = "market", nullable = false, length = 20)
     private String market;
 
-    /** The symbol. */
-    @Column(name = "symbol", nullable = false, length = 20)
-    private String symbol;
-
     /** The name. */
     @Column(name = "name", length = 100)
     private String name;
 
-    /** The trade type. */
-    @Column(name = "trade_type", nullable = false, length = 10)
-    @Enumerated(EnumType.STRING)
-    private TradeType tradeType;
-
-    /** The quantity. */
-    @Column(name = "quantity", nullable = false, precision = 19, scale = 4)
-    private BigDecimal quantity;
-
-    /** The price. */
-    @Column(name = "price", nullable = false, precision = 19, scale = 4)
-    private BigDecimal price;
-
-    /** The amount. */
-    @Column(name = "amount", nullable = false, precision = 19, scale = 4)
-    private BigDecimal amount;
-
-    /** The trade time. */
-    @Column(name = "trade_time", nullable = false)
-    private LocalDateTime tradeTime;
-
-    /** The created at timestamp. */
-    @CreationTimestamp
-    @Column(name = "created_at", updatable = false)
-    @Setter(AccessLevel.NONE)
-    private LocalDateTime createdAt;
-
     /** The status. */
     @Column(name = "status", length = 20)
     @Enumerated(EnumType.STRING)
-    @Builder.Default
-    private TradeStatus status = TradeStatus.SUCCESS;
+    private TradeStatus status;
 
     /** The notes. */
     @Column(name = "notes", length = 500)
     private String notes;
-
-    /**
-     * Trade status enum.
-     */
-    public enum TradeStatus {
-
-        /** Pending status. */
-        PENDING,
-
-        /** Success status. */
-        SUCCESS,
-
-        /** Failed status. */
-        FAILED,
-
-        /** Cancelled status. */
-        CANCELLED
-    }
 }
