+++ koduck-backend/src/main/java/com/koduck/entity/enums/TradeStatus.java
@@ -1,0 +1,22 @@
+package com.koduck.entity.enums;
+
+/**
+ * Trade status enumeration.
+ * Extracted from Trade.TradeStatus to eliminate inner enum redundancy.
+ *
+ * @author Koduck Team
+ */
+public enum TradeStatus {
+
+    /** Pending status. */
+    PENDING,
+
+    /** Success status. */
+    SUCCESS,
+
+    /** Failed status. */
+    FAILED,
+
+    /** Cancelled status. */
+    CANCELLED
+}
